Move utility function to utils file, add doc comment, update usage
